refactor: organize components into proper directories (Phase 1)
- Create layout/ directory for navigation and structural components
- Create utils/ directory for utility components
- Move folder-top-bar.tsx to dashboard/ (dashboard-specific)
- Move header.tsx and footer.tsx to layout/
- Move posthog-provider.tsx to utils/
- Update all import paths to reflect new structure
- Fix relative imports in header.tsx to use absolute paths

This is Phase 1 of component organization - low risk, high impact moves
that improve component discoverability and follow established patterns.
