+++ src/app/tests/layout.tsx
@@ -2,7 +2,7 @@
 
 import "@/app/[lang]/globals.css";
 import { SessionProvider } from "next-auth/react";
-import { PostHogProvider } from "@/components/posthog-provider";
+import { PostHogProvider } from "@/components/utils/posthog-provider";
 
 export default function TestsLayout({ children }: { children: React.ReactNode }) {
   return (
